factor data panel into its own file
